requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
streamlit>=1.28.0
chromadb>=0.4.0
langchain-community>=0.0.20
langgraph>=0.1.0
